+++ arangod/VectorIndex/VectorIndexBuildManager.cpp
@@ -157,7 +157,7 @@ void VectorIndexBuildManager::fulfillAllWaiters(Result const& result) {
 
 bool VectorIndexBuildManager::shouldSkipRetry(
     FailedBuildsMap const& failedBuilds, std::uint64_t objectId,
-    std::int64_t currentDocCount) {
+    std::uint64_t currentDocCount) {
   auto const it = failedBuilds.find(objectId);
   if (it == failedBuilds.end()) {
     return false;
@@ -247,8 +247,7 @@ void VectorIndexBuildManager::scanAndBuild(std::stop_token const& stopToken,
 
         auto const* rcoll =
             static_cast<RocksDBCollection*>(coll->getPhysical());
-        auto const numDocs =
-            static_cast<std::int64_t>(rcoll->meta().numberDocuments());
+        auto const numDocs = rcoll->meta().numberDocuments();
         if (numDocs < vecIdx.trainingThreshold()) {
           skippedWaiters.insert(vecIdx.id().id());
           reportIndexError(
